Move euth permissions tests to workflows module
