+++ src/adhocracy_euth/adhocracy_euth/workflows/test_collaborative_text/resources.json
